get.py

+++ get.py
@@ -4,6 +4,7 @@ import os
 import json
 
 
+
 def get_word_epidemic(date):
     """
     获取世界疫情数据
@@ -15,9 +16,10 @@ def get_word_epidemic(date):
         data = pd.read_csv(open(
             os.path.join("data/world-epidemic/", date_name + ".csv"),
             'r', encoding="utf-8"), dtype=np.object)
-        confirmed = data["Confirmed"]
+        confirmed = data["Confirmed"].astype(int)
         max = confirmed.drop(0).max()
         min = data["Confirmed"].min()
+
         result_list = []
         for index in range(0, data.shape[0]):
             item = data.iloc[index].tolist()
@@ -26,6 +28,7 @@ def get_word_epidemic(date):
     else:
         print("没有这一天的数据！")
         return None,None,None
+
 
 
 def get_country_epidemic(date, country):
@@ -42,7 +45,7 @@ def get_country_epidemic(date, country):
         data = pd.read_csv(
             open(os.path.join("data/country-epidemic/", date_name + "-" + country + ".csv"),
                  "r", encoding="utf-8"), dtype=np.object)
-        confirmed = data["Confirmed"]
+        confirmed = data["Confirmed"].astype(int)
         max = confirmed.max()
         min = confirmed.min()
         result = []
@@ -52,7 +55,6 @@ def get_country_epidemic(date, country):
     else:
         print("没有相关数据！")
         return None
-
 
 def get_country_epidemic_summary(country):
     """
